// src/types.ts
export enum ButtonLabels {
  ZERO,
  ONE,
  TWO,
  THREE,
  FOUR,
  FIVE,
  SIX,
  SEVEN,
  EIGHT,
  NINE,
  ADDITION = '+',
  SUBTRACTION = '-',
  MULTIPLICATION = '*',
  DIVISION = '/',
  EQUAL = '=',
  CLEAR = 'C'
}

export enum ButtonTypes {
  NUMBERS = 'numbers',
  OPERATORS = 'operators',
  EQUAL = 'equal',
  CLEAR = 'clear'
}